Bring over Erlang frontend from Lumen
- Provides a diagnostics library for debug info and source spans
- Provides an Erlang frontend/parser built on LALRPOP
    - Supports all valid Erlang syntax
    - Supports all standard attributes (e.g. compiler directives,
    typespecs/callbacks, etc.)
    - Supports preprocessor directives/macro expansion

NOTE: This **does not** support parse transforms (yet). That is a
difficult topic, since it requires evaluation of Erlang source code.
While the preprocessor supports some degree of partial evaluation, it is
limited to the things supported in preprocessor directives, whereas
parse transforms require support for significantly more functionality,
possibly the full language.