Refactor diagnostics
This changset reworks diagnostics to use codespan/codespan-reporting
for the underlying primitives and reporting functionality, but providing
our own implementation of the Files trait so that we handle highly
concurrent cases better.

The CodeMap type is designed to be allocated in an Arc and then cloned
freely to pass around, and its use should be primarily to add files and
look them up from a source value. Internally it uses two maps, one that
holds the actual sources, and one that tracks paths which have been
loaded to avoid loading duplicates into memory.

The SourceFile type is pretty similar to the old FileMap type, but one
of the key differences now is the use of a SourceId to identify sources
uniquely. This is how lookups in the CodeMap are performed.

One of the big differences is in how we look up files from a source
index or span.  The old CodeMap worked by searching all files for the one
which contained a given ByteIndex, but we don't operate that way now. In
order to accomodate this cleanly, there are some new types to represent
spans and indices at a high-level, the SourceSpan and SourceIndex types.
The SourceIndex type is a 64-bit value that uses half of the bits for
the SourceId, and half of the bits for the actual ByteIndex; SourceSpans
decompose two SourceIndex values into a single SourceId and two
ByteIndex values. The gist is that both types can identify what SourceId
they originate from, which can then be used to look up files. The most
pressing need for this is in the Diagnostics API exposed by
codespan-reporting, since its Label type requires the SourceId, not just
any index or span. The changes described here are what produced the bulk of
the diff in this patch.

I had originally wanted libeir_* to avoid having a concrete
representation like CodeMap/SourceFile at all, and entirely build on
traits and types provided by codespan/codespan-reporting, but
unfortunately there isn't a clean way to do that right now. I think it
makes sense to move that way eventually, but to do so, we'll need to sit
down and think about how we're using diagnostics in these crates.