Make `Case` operation a `DynOp`
This makes the `Case` operation into a `DynOp`. This has the effect of
completely decoupling the `Case` operation from the core EIR IR
implementation. This is desirable because the `Case` operation is only
ever used in the early stages of the IR, before pattern match
compilation is preformed.

Changes includes:
* Removing `Case` from the Op variant
* Add a new `Case` module with the `DynOp` implementation
* There is now one separate `PatternContainer` for each `Case`
* Move frontends to use the new `Case` `DynOp`
* Make pattern match compilation now uses the new `Case` `DynOp`