Fix issue in graph similification pass
The graph simplification pass has several strategies it can use to
perform the simplification. Which strategy is used in which case is
determined by how the source IR looks in each case.

Here, a bug in the checks for one of the strategies (terminating chain
strategy) caused it to be applied in cases where it shouldn't be.

The fix was to only run the strategy if no dynamic values from the chain
are used after the target block.